#55 add many-to-many between order and menuItem and order has a list of...

+++ Project/Backend/src/main/java/rhul/cs2810/repository/MenuItemRepository.java
@@ -2,13 +2,12 @@ package rhul.cs2810.repository;
 
 
 import org.springframework.data.repository.CrudRepository;
-
 import rhul.cs2810.model.MenuItem;
 
 /**
  * Repository for managing Menu Item entities.
  */
-public interface MenuItemRepository extends CrudRepository<MenuItem, Long> {
+public interface MenuItemRepository extends CrudRepository<MenuItem, Integer> {
 
 }
