+++ Project/Backend/src/main/java/rhul/cs2810/repository/OrderRepository.java
@@ -1,13 +1,15 @@
 package rhul.cs2810.repository;
 
+
+
 // might not be needed? orders stored in customer class
 
 import org.springframework.data.repository.CrudRepository;
-import jakarta.persistence.criteria.Order;
+import rhul.cs2810.model.Order;
 
 /**
  * Repository for managing Order entities.
  */
-public interface OrderRepository extends CrudRepository<Order, Long> {
+public interface OrderRepository extends CrudRepository<Order, Integer> {
 
 }
